Implement OpendalAdapter
