+++ jumpstarter/client/adapters/__init__.py
@@ -1,5 +1,6 @@
 from .common import ClientAdapter
 from .novnc import NovncAdapter
+from .opendal import OpendalAdapter
 from .portforward import PortforwardAdapter
 
-__all__ = ["ClientAdapter", "PortforwardAdapter", "NovncAdapter"]
+__all__ = ["ClientAdapter", "PortforwardAdapter", "NovncAdapter", "OpendalAdapter"]
