+++ jumpstarter/client/mixins.py
@@ -5,7 +5,6 @@ Mixins for extending DriverClient
 import socket
 from contextlib import contextmanager
 
-from opendal import Operator
 from pexpect.fdpexpect import fdspawn
 
 from jumpstarter.client.adapters import PortforwardAdapter
@@ -23,12 +22,3 @@ class ExpectMixin:
             with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
                 s.connect(addr)
                 yield fdspawn(s)
-
-
-class ResourceMixin:
-    """Resource"""
-
-    @contextmanager
-    def file(self, operator: Operator, path: str):
-        with self.portal.wrap_async_context_manager(self.file_async(operator.to_async_operator(), path)) as uuid:
-            yield uuid
